+++ TelegramBotAPI-extensions-utils/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/extensions/utils/CommonMessageConversations.kt
@@ -1,7 +1,6 @@
 package com.github.insanusmokrassar.TelegramBotAPI.extensions.utils
 
-import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.CommonMessage
-import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.ContentMessage
+import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.*
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.content.abstracts.MessageContent
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.content.abstracts.PossiblySentViaBotCommonMessage
 import kotlinx.coroutines.flow.*
@@ -12,16 +11,20 @@ import kotlinx.coroutines.flow.*
 fun <C: MessageContent, T : ContentMessage<C>> Flow<T>.onlyCommonMessages() = filterIsInstance<CommonMessage<C>>()
 
 /**
+ * Shortcut for [onlyCommonMessages]
+ */
+@Suppress("NOTHING_TO_INLINE")
+inline fun <C: MessageContent, T : ContentMessage<C>> Flow<T>.commonMessages() = onlyCommonMessages()
+
+/**
  * Filter the messages and checking that incoming [CommonMessage] is [PossiblySentViaBotCommonMessage] and its
  * [PossiblySentViaBotCommonMessage.senderBot] is not null
  */
-fun <T : MessageContent> Flow<CommonMessage<T>>.onlySentViaBot() = mapNotNull {
-    (it as? PossiblySentViaBotCommonMessage) ?.let { possiblySentViaBot ->
-        if (possiblySentViaBot.senderBot != null) {
-            possiblySentViaBot
-        } else {
-            null
-        }
+fun <MC : MessageContent, M : ContentMessage<MC>> Flow<M>.onlySentViaBot() = mapNotNull {
+    if (it is PossiblySentViaBot && it.senderBot != null) {
+        it
+    } else {
+        null
     }
 }
 
@@ -29,6 +32,6 @@ fun <T : MessageContent> Flow<CommonMessage<T>>.onlySentViaBot() = mapNotNull {
  * Filter the messages and checking that incoming [CommonMessage] not is [PossiblySentViaBotCommonMessage] or its
  * [PossiblySentViaBotCommonMessage.senderBot] is null
  */
-fun <T : MessageContent> Flow<CommonMessage<T>>.withoutSentViaBot() = filter {
-    it !is PossiblySentViaBotCommonMessage || it.senderBot == null
+fun <MC : MessageContent, M : ContentMessage<MC>> Flow<M>.withoutSentViaBot() = filter {
+    it !is PossiblySentViaBot || it.senderBot == null
 }