+++ CHANGELOG.md
@@ -53,6 +53,19 @@
         * `handleSafely` has changed its signature
         * `executeUnsafe` has changed its signature
 
+### 0.28.2
+
+* `TelegramBotAPI-extensions-utils`:
+    * Several commands shortcuts for `Flow<ContentMessage<TextContent>>` has been added:
+        * `filterExactCommands`
+        * `filterCommandsInsideTextMessages`
+        * `filterCommandsWithArgs`
+    * Extension `Flow<BaseSentMessageUpdate>.filterCommandsWithArgs` has changed its signature: now it will also have
+    original message paired with list of text sources
+    * Shortcut method `commonMessages` for `onlyCommonMessages`
+    * Shortcuts `onlySentViaBot` and `withoutSentViaBot` now are extensions for any `Flow` with types which implementing
+    `ContentMessage`
+
 ## 0.27.0
 
 * `Common`: